Abstract the tECDSA signing announcer logic
Here we abstract the tECDSA signing announcer and move it to the
`pkg/protocol/announcer` package. This logic is generic and can be used by
multiple interactive protocols.
