+++ pkg/tbtc/gen/pb/message.proto
@@ -8,9 +8,3 @@ message StopPill {
     string dkgSeed = 2;
     string messageToSign = 3;
 }
-
-message SigningAnnouncementMessage {
-    uint32 senderID = 1;
-    bytes message = 2;
-    uint64 attemptNumber = 3;
-}
